+++ tests/__init__.py
@@ -1,7 +1,7 @@
 """
 tests package.
 
-Test suite for the loggingx library.
+Test suite for the logenrich library.
 
 Author: Ronaldo Webb
 Since: 1.0.0